pyproject.toml
[project]
name = "paper-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "crewai>=1.9.3",
    "langchain-anthropic>=0.3.22",
    "langchain-openai>=0.3.23",
    "markitdown>=0.1.4",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
